requirements.txt
fastapi==0.115.6
mediapipe==0.10.20
numpy==1.26.0
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
Pillow==11.1.0
